+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 XCommon CMake Change Log
 ========================
 
+UNRELEASED
+----------
+
+  * ADDED:  "Depends on" column in generated manifest file (when FULL_MANIFEST=TRUE)
+  * CHANGED: Removed "Dependency_requirement" column in generated manifest file (when FULL_MANIFEST=TRUE)
+
 1.0.0
 -----